CLI: Register the TypeScript loader in-thread on modern Node
The worker-thread loader (module.register) makes every module load in
the process — plain JS included — pay a synchronous IPC round-trip to
the hook worker, which profiles at 100-200ms per CLI invocation. On Node
with module.registerHooks (22.15+) and module.stripTypeScriptTypes
(22.13+) the TypeScript transform now runs synchronously in-process
instead, eliminating that tax; older Node keeps the worker loader as a
fallback.

The pure helpers (extension resolution, relative-import rewriting) move
to bin/loader-utils.ts, shared by both registration paths.
